テストしやすいように main を分割
