--- ch11/ex01/charcount.go
@@ -1,45 +1,0 @@
-// charcount は Unicode 文字の数を計算します
-package main
-
-import (
-	"bufio"
-	"fmt"
-	"io"
-	"os"
-	"unicode"
-	"unicode/utf8"
-)
-
-func main() {
-	counts := make(map[rune]int)
-	var utflen [utf8.UTFMax + 1]int
-	invalid := 0
-
-	in := bufio.NewReader(os.Stdin)
-	for {
-		r, n, err := in.ReadRune() // rune, nbytes, error を返す
-		if err == io.EOF {
-			break
-		}
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "charcount: %v\n", err)
-			os.Exit(1)
-		}
-		if r == unicode.ReplacementChar && n == 1 {
-			invalid++
-			continue
-		}
-		counts[r]++
-		utflen[n]++
-	}
-	fmt.Printf("rune\tconut\n")
-	for c, n := range counts {
-		fmt.Printf("%q\t%d\n", c, n)
-	}
-	fmt.Printf("\nlen\tconut\n")
-	for i, n := range utflen {
-		if i > 0 {
-			fmt.Printf("%d\t%d\n", i, n)
-		}
-	}
-}
